+++ src/main/java/org/example/repositories/AgendaRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AgendaRepository extends JpaRepository<Agenda, Integer> {}
+public interface AgendaRepository extends JpaRepository<Agenda, Long> {}
