+++ src/main/java/org/example/repositories/EnderecoRepository.java
@@ -3,8 +3,7 @@ package org.example.repositories;
 import org.example.models.Endereco;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import java.util.ArrayList;
-import java.util.List;
+
 
 @Repository
-public interface EnderecoRepository extends JpaRepository<Endereco, Integer> {}
+public interface EnderecoRepository extends JpaRepository<Endereco, Long> {}
