+++ src/main/java/org/example/repositories/FuncionarioRepository.java
@@ -4,8 +4,5 @@ import org.example.models.Funcionario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.ArrayList;
-import java.util.List;
-
 @Repository
-public interface FuncionarioRepository extends JpaRepository<Funcionario, Integer> {}
+public interface FuncionarioRepository extends JpaRepository<Funcionario, Long> {}
