+++ src/main/java/org/example/repositories/ServicoRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ServicoRepository extends JpaRepository<Servico, Integer> {}
+public interface ServicoRepository extends JpaRepository<Servico, Long> {}
