+++ src/main/java/org/example/repositories/TipoAcessoRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoAcessoRepository extends JpaRepository<TipoAcesso, Integer> {}
+public interface TipoAcessoRepository extends JpaRepository<TipoAcesso, Long> {}
